added things but nothing works

+++ src/main/java/com/bmd_app/bmd_app/Repository/DeliveryRepository.java
@@ -3,5 +3,5 @@ package com.bmd_app.bmd_app.Repository;
 import com.bmd_app.bmd_app.Entity.Delivery;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DeliveryRepository  extends CrudRepository<Delivery, Integer> {
+public interface DeliveryRepository  extends CrudRepository<Delivery, Long> {
 }
