+++ src/main/java/com/bmd_app/bmd_app/Repository/RequestRepository.java
@@ -3,5 +3,5 @@ package com.bmd_app.bmd_app.Repository;
 import com.bmd_app.bmd_app.Entity.Request;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RequestRepository extends CrudRepository<Request, Integer> {
+public interface RequestRepository extends CrudRepository<Request, Long> {
 }
